structure/code refactor of the address_profile changes according with project policies
