+++ lib/pages/customer_user/profile_addresses/confirm_address.dart
@@ -3,8 +3,10 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
 import 'package:spl_front/bloc/ui_management/search_places/search_places_bloc.dart';
+import 'package:spl_front/utils/strings/address_strings.dart';
+import 'package:spl_front/widgets/addresses/helpers/address_dialogs.dart';
 
-import '../../bloc/ui_management/address/address_bloc.dart';
+import '../../../bloc/ui_management/address/address_bloc.dart';
 
 class ConfirmAddressPage extends StatefulWidget {
   const ConfirmAddressPage({super.key});
@@ -26,7 +28,7 @@ class _ConfirmAddressPageState extends State<ConfirmAddressPage> {
         return Scaffold(
           appBar: AppBar(
             title: Text(
-              'Detalles de La Dirección',
+              AddressStrings.addressDetails,
               style: TextStyle(
                   color: Colors.black,
                   fontSize: 20,
@@ -45,7 +47,7 @@ class _ConfirmAddressPageState extends State<ConfirmAddressPage> {
               crossAxisAlignment: CrossAxisAlignment.start,
               children: [
                 Text(
-                  'Dirección',
+                  AddressStrings.address,
                   style: TextStyle(fontSize: 18, fontWeight: FontWeight.bold),
                 ),
                 Text(
@@ -56,7 +58,7 @@ class _ConfirmAddressPageState extends State<ConfirmAddressPage> {
                 ),
                 SizedBox(height: 15),
 
-                // Mapa (Estático con el icono del marcador)
+                // Static Map with the selected location
                 Container(
                   height: 200,
                   decoration: BoxDecoration(
@@ -64,7 +66,7 @@ class _ConfirmAddressPageState extends State<ConfirmAddressPage> {
                     border: Border.all(color: Colors.grey),
                     boxShadow: [
                       BoxShadow(
-                        color: Colors.grey.withOpacity(0.2),
+                        color: Colors.grey,
                         blurRadius: 4,
                         spreadRadius: 2,
                       ),
@@ -94,7 +96,7 @@ class _ConfirmAddressPageState extends State<ConfirmAddressPage> {
                             ),
                           ),
                           child: Text(
-                            'Ajustar',
+                            AddressStrings.adjustLocation,
                             style: TextStyle(color: Colors.white),
                           ),
                         ),
@@ -103,18 +105,18 @@ class _ConfirmAddressPageState extends State<ConfirmAddressPage> {
                   ),
                 ),
 
-                // Etiqueta editable
+                // Editable label
                 SizedBox(height: 25),
                 Text(
-                  'Etiqueta tu dirección',
+                  AddressStrings.labelAddressInstruction,
                   style: TextStyle(fontSize: 16, fontWeight: FontWeight.bold),
                 ),
                 SizedBox(height: 10),
                 TextFormField(
                   controller: labelController,
                   decoration: InputDecoration(
-                    labelText: 'Etiqueta tu dirección',
-                    hintText: 'Etiqueta (p.e. Casa, Oficina)',
+                    labelText: AddressStrings.labelAddress,
+                    hintText: AddressStrings.hintAddress,
                     labelStyle: TextStyle(color: Colors.black38),
                     filled: true,
                     fillColor: Colors.grey.shade200,
@@ -135,16 +137,16 @@ class _ConfirmAddressPageState extends State<ConfirmAddressPage> {
                 // Detalles de la dirección
                 SizedBox(height: 20),
                 Text(
-                  'Detalles de tu dirección',
+                  AddressStrings.detailsAddressInstruction,
                   style: TextStyle(fontSize: 16, fontWeight: FontWeight.bold),
                 ),
                 SizedBox(height: 10),
                 TextFormField(
                   controller: detailsController,
                   decoration: InputDecoration(
-                    labelText: 'Detalles de tu dirección',
+                    labelText: AddressStrings.labelDetails,
                     labelStyle: TextStyle(color: Colors.black38),
-                    hintText: 'Número de Apartamento u oficina',
+                    hintText: AddressStrings.hintDetails,
                     filled: true,
                     fillColor: Colors.grey.shade200,
                     border: OutlineInputBorder(),
@@ -161,17 +163,14 @@ class _ConfirmAddressPageState extends State<ConfirmAddressPage> {
                   ),
                 ),
 
-                // Empty space before button
-                SizedBox(height: 40), // Adds space before button
-
-                // Centering the "Guardar Dirección" button at the bottom
+                SizedBox(height: 40),
                 Center(
                   child: ElevatedButton(
                     onPressed: () {
                       if (labelController.text.isEmpty ||
                           detailsController.text.isEmpty) {
-                        // Mostrar un dialogo si alguno de los campos está vacío
-                        _showErrorDialog(context);
+                        // Show a dialog if any of the fields is empty
+                        showErrorDialog(context);
                       } else {
                         final addressBloc =
                             BlocProvider.of<AddressBloc>(context);
@@ -200,77 +199,13 @@ class _ConfirmAddressPageState extends State<ConfirmAddressPage> {
                         borderRadius: BorderRadius.circular(8),
                       ),
                     ),
-                    child: Text('Guardar Dirección',
+                    child: Text(AddressStrings.saveAddress,
                         style: TextStyle(color: Colors.white)),
                   ),
                 ),
               ],
             ),
           ),
-        );
-      },
-    );
-  }
-
-  void _showErrorDialog(BuildContext context) {
-    double mediaQueryWidth = (MediaQuery.of(context).size.width / 1.5);
-
-    showDialog(
-      context: context,
-      builder: (context) {
-        return AlertDialog(
-          shape: RoundedRectangleBorder(
-            borderRadius: BorderRadius.circular(12), // Round the corners
-          ),
-          title: Align(
-            alignment: Alignment.centerLeft, // Align title to the left
-            child: Text(
-              'Campos Obligatorios',
-              style: TextStyle(
-                color: Colors.black, // Black color for the title
-                fontWeight: FontWeight.bold, // Bold font for the title
-                fontSize: 20, // Larger font size for the title
-              ),
-            ),
-          ),
-          content: Padding(
-            padding: const EdgeInsets.symmetric(vertical: 10.0),
-            child: Text(
-              maxLines: 2,
-              overflow: TextOverflow.ellipsis,
-              'El nombre y los detalles de la ubicación son obligatorios.',
-              style: TextStyle(
-                color:
-                    Colors.black87, // Slightly lighter black color for content
-                fontSize: 16, // Larger font size for content
-                fontWeight: FontWeight.w400, // Regular weight for the content
-              ),
-              textAlign: TextAlign.left, // Align content to the left
-            ),
-          ),
-          actions: [
-            // Button styled similar to the GPS dialog
-            TextButton(
-              style: TextButton.styleFrom(
-                minimumSize: Size(mediaQueryWidth, 50),
-                backgroundColor: Colors.blue,
-                shape: RoundedRectangleBorder(
-                  borderRadius: BorderRadius.circular(8),
-                ),
-              ),
-              onPressed: () {
-                Navigator.pop(context);
-              },
-              child: Text(
-                'Aceptar',
-                style: TextStyle(
-                  color: Colors.white,
-                  fontWeight: FontWeight.bold,
-                  fontSize: 16, // Slightly larger font for the button text
-                ),
-              ),
-            ),
-          ],
         );
       },
     );
