+++ lib/pages/customer_user/profile_addresses/map_address_page.dart
@@ -4,8 +4,8 @@ import 'package:google_maps_flutter/google_maps_flutter.dart';
 import 'package:spl_front/bloc/ui_management/location/location_bloc.dart';
 import 'package:spl_front/bloc/ui_management/map/map_bloc.dart';
 import 'package:spl_front/bloc/ui_management/search_places/search_places_bloc.dart';
-import 'package:spl_front/widgets/map/helpers/buttons/back_widget.dart';
-import 'package:spl_front/widgets/map/helpers/buttons/follow_user_widget.dart';
+import 'package:spl_front/widgets/buttons/back_widget.dart';
+import 'package:spl_front/widgets/buttons/follow_user_widget.dart';
 import 'package:spl_front/widgets/map/manual_marker.dart';
 import 'package:spl_front/widgets/map/map_view_address.dart';
