+++ lib/routes/routes.dart
@@ -4,10 +4,10 @@ import 'package:spl_front/pages/business_user/add_product.dart';
 import 'package:spl_front/pages/business_user/chats_business_user.dart';
 import 'package:spl_front/pages/business_user/dashboard_business_user.dart';
 import 'package:spl_front/pages/business_user/profile_business_user.dart';
-import 'package:spl_front/pages/customer_user/add_address.dart';
-import 'package:spl_front/pages/customer_user/confirm_address.dart';
 import 'package:spl_front/pages/customer_user/dashboard_customer_user.dart';
-import 'package:spl_front/pages/customer_user/map_address_page.dart';
+import 'package:spl_front/pages/customer_user/profile_addresses/add_address.dart';
+import 'package:spl_front/pages/customer_user/profile_addresses/confirm_address.dart';
+import 'package:spl_front/pages/customer_user/profile_addresses/map_address_page.dart';
 import 'package:spl_front/pages/customer_user/profile_customer_user.dart';
 import 'package:spl_front/pages/delivery_user/profile_delivery.dart';
 import 'package:spl_front/pages/login_page_web.dart';
