+++ lib/utils/strings/address_strings.dart
@@ -1,0 +1,23 @@
+class AddressStrings {
+  static const addressDetails = 'Detalles de La Dirección';
+  static const address = 'Dirección';
+  static const labelAddressInstruction = 'Etiqueta tu dirección';
+  static const labelAddress = 'Etiqueta tu Dirección';
+  static const hintAddress = 'Etiqueta (p.e. Casa, Oficina)';
+  static const detailsAddressInstruction = 'Detalles de tu dirección';
+  static const labelDetails = 'Detalles de tu dirección';
+  static const hintDetails = 'Detalles (p.e. Número de Apartamento u Oficina)';
+  static const saveAddress = 'Guardar Dirección';
+  static const mandatoryFields = 'Campos Obligatorios';
+  static const mandatoryFieldsDescription =
+      'El nombre y los detalles de la ubicación son obligatorios.';
+  static const accept = 'Aceptar';
+  static const addNewAddress = 'Agrega una Nueva Dirección';
+  static const addAddress = 'Agregar Dirección';
+  static const searchResults = 'Resultados de la Búsqueda';
+  static const adjustLocation = 'Ajustar';
+  static const deliveryAddresses = 'Direcciones de Entrega';
+  static const inputLabelName = 'Nombre';
+  static const inputLabelDetails = 'Detalles';
+  static const searchAddress = 'Busca tu dirección';
+}
