+++ lib/utils/strings/map_strings.dart
@@ -1,0 +1,13 @@
+class MapStrings {
+  static const selectOnMap = 'Seleccionar en el Mapa';
+  static const destinationConfirm = 'Confirmar Destino';
+  static const activeGPS = 'Activar GPS';
+  static const activeGPSDescription =
+      'Por favor, active el GPS para seleccionar la dirección en el mapa.';
+  static const cancel = 'Cancelar';
+
+  static const locationPermission = 'Permisos de Ubicación';
+  static const locationPermissionDescription =
+      'Por favor, otorgue los permisos para acceder a la funcionalidad del Mapa.';
+  static const askLocationPermission = 'Solicitar Permiso';
+}
