+++ lib/widgets/addresses/helpers/debouncer.dart
@@ -1,0 +1,18 @@
+import 'dart:async';
+
+import 'package:flutter/material.dart';
+
+class Debouncer {
+  final Duration duration;
+  VoidCallback? action;
+  Timer? _timer;
+
+  Debouncer({required this.duration});
+
+  run(VoidCallback action) {
+    if (_timer != null) {
+      _timer!.cancel();
+    }
+    _timer = Timer(duration, action);
+  }
+}
