+++ lib/widgets/map/manual_marker.dart
@@ -2,6 +2,7 @@ import 'package:animate_do/animate_do.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:spl_front/bloc/ui_management/search_places/search_places_bloc.dart';
+import 'package:spl_front/utils/strings/map_strings.dart';
 
 import '../../bloc/ui_management/map/map_bloc.dart';
 
@@ -50,7 +51,7 @@ class ManualMarker extends StatelessWidget {
                 minWidth: size.width - 120,
                 color: Colors.black,
                 child: const Text(
-                  'Confirmar Destino',
+                  MapStrings.destinationConfirm,
                   style: TextStyle(color: Colors.white),
                 ),
               ),
